chore: clean up all compiler and clippy warnings
- Remove unused imports (MerkleDiffResult, craftobj_transfer, HashMap, error, InboundMessage, write_response_frame)
- Remove dead code (compute_cid_health, load_own_pieces, request_piece methods)
- Fix unused variables with _ prefix where values are captured but not yet used
- Remove unnecessary mut qualifiers
- Remove unreachable match arms (single-variant enum)
- Replace filter_map(|line| line.ok()) with map_while(Result::ok)
- Fix redundant closures in craftobj-transfer (16 fixes via clippy --fix)
- Fix explicit counter loop (seq_id) in piece_transfer
- Remove unused struct fields (provider_regions, local_region in ConnectionPool)
- Add targeted #[allow] for architectural warnings (too_many_arguments, large_enum_variant, etc.)

Result: zero warnings on cargo build, cargo clippy, and cargo test --lib

+++ crates/daemon/src/disk_monitor.rs
@@ -37,9 +37,9 @@ pub fn get_disk_space(path: &Path) -> Option<DiskSpace> {
             return None;
         }
         let stat = unsafe { stat.assume_init() };
-        let total = stat.f_blocks as u64 * stat.f_frsize as u64;
-        let available = stat.f_bavail as u64 * stat.f_frsize as u64;
-        let used = total.saturating_sub(stat.f_bfree as u64 * stat.f_frsize as u64);
+        let total = stat.f_blocks as u64 * stat.f_frsize;
+        let available = stat.f_bavail as u64 * stat.f_frsize;
+        let used = total.saturating_sub(stat.f_bfree as u64 * stat.f_frsize);
         let percent = if total > 0 { used as f64 / total as f64 } else { 0.0 };
         Some(DiskSpace {
             used_bytes: used,
